cgroups2/Stats: split out reading cpu, memory, memory events
As outlined in the PR, usage_usec, user_usec, and system_use may
always exist, but in situations where the files parsed do not
exist, we shouldn't have to try to assign values (which would always
be 0).

Split out parsing of these to separate functions, so make this more
transparent.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>